requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0

apscheduler>=3.10.4

python-telegram-bot >= 20.0

requests>=2.31.0

jsonschema>=4.19.0

pyTelegramBotAPI